Refactor finding audit detail to SFC
